libnvme: move accessors.[ch] and accessors-fabrics.[ch] to generated subdir
keep generated files out of src/nvme/ so they're visually separate
from the hand-written files around them.

Also fixes stale-header fallback bugs this move would otherwise cause:
tid.c/tid.h, exclusion.c, and discoverd/src/tid.h all reference these
headers via the public "nvme/accessors-fabrics.h" path rather than a
same-directory include, which was silently resolving to a stale copy
in /usr/include from a previous install once the real file moved.

Also fixes a -Werror build break introduced by the generated/ subdir
move: nvme_incdir (added in the previous commit to let a quoted
#include "private.h" resolve from nvme/generated/*.c) put
-Ilibnvme/src/nvme on every compile command, which made
libnvme/src/nvme/endian.h's own #include <endian.h> resolve back to
itself instead of glibc's real header -- #pragma once silently
no-ops the self-reference, dropping the declarations for
htole64/htobe64/etc. Only visible as a warning locally; CI builds
with --werror, where it's fatal. Removes nvme_incdir and instead
makes both generators emit correct explicit relative includes
(../private.h, ../private-fabrics.h) for the headers they parsed,
now that their .c output lives one directory below. Verified with
--werror (default and -Dfabrics=disabled), full test suite, and a
live smoke test.
