+++ discoverd/src/tid.h
@@ -9,7 +9,7 @@
 
 #include <stdbool.h>
 
-#include <nvme/accessors-fabrics.h>
+#include <nvme/generated/accessors-fabrics.h>
 #include <nvme/tid.h>
 
 #include <string-util.h>
