+++ libnvme/src/fabrics-includes.h.in
@@ -1,4 +1,4 @@
-#include <nvme/accessors-fabrics.h>
+#include <nvme/generated/accessors-fabrics.h>
 #include <nvme/config.h>
 #include <nvme/crypto.h>
 #include <nvme/exclusion.h>
