+++ libnvme/src/nvme/tid.c
@@ -15,7 +15,7 @@
 #include <compiler-attributes.h>
 #include <string-util.h>
 
-#include "nvme/accessors-fabrics.h"
+#include "nvme/generated/accessors-fabrics.h"
 #include "nvme/lib.h"
 #include "nvme/tid.h"
 #include "private-fabrics.h"
